requirements.txt
pyquery
pyyaml
requests
